Make people rich.

+++ src/main/java/joshie/harvestmoon/commands/CommandGold.java
@@ -19,7 +19,7 @@ public class CommandGold extends CommandBase {
     public void processCommand(ICommandSender sender, String[] parameters) {
         if (parameters == null || parameters.length < 1 || parameters.length > 2) return;
         try {
-            int amount = parameters.length == 1 ? Integer.parseInt(parameters[0]) : Integer.parseInt(parameters[1]);
+            long amount = parameters.length == 1 ? Long.parseLong(parameters[0]) : Long.parseLong(parameters[1]);
             boolean set = parameters.length == 1 || parameters[0].equals("add") ? false : true;
             PacketHandler.sendToServer(new PacketGoldCommand(amount, set));
         } catch (NumberFormatException e) {}
