+++ src/main/java/joshie/harvestmoon/util/SellStack.java
@@ -4,20 +4,20 @@ import net.minecraft.item.ItemStack;
 
 public class SellStack extends SafeStack {
     public int amount;
-    public int sell;
+    public long sell;
 
-    public SellStack(String item, int damage, int amount, int sell) {
+    public SellStack(String item, int damage, int amount, long sell) {
         super(item, damage);
         this.amount = amount;
         this.sell = sell;
     }
     
-    public SellStack(ItemStack stack, int sell) {
+    public SellStack(ItemStack stack, long sell) {
         super(stack);
         this.sell = sell;
     }
     
-    public void add(int sell) {
+    public void add(long sell) {
         this.sell += sell;
         amount++;
     }
